Week 15 Checkpoint
TEAM-EVO-SPORTS-EMPORIUM

Features
- 3d animation of shoe spinning
- purchasing of items
- like items 
- checkout and purchase

Technologies
HTML will be the base of our website. While CSS & Jquery will be used for the raw design of the website. Restdb will be deployed as database storage for our website, to get feedback and storing of information. Lottie was used it for our startup page animation.